% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSurveyData.R
\name{cleanSurveyData}
\alias{cleanSurveyData}
\title{Clean Survey Data}
\arguments{
\item{data}{A data frame containing survey data.}

\item{options}{A list of options specifying how to handle missing data and recoding of responses.}
}
\value{
A data frame with cleaned and preprocessed data.
}
\description{
Clean Survey Data
}
\examples{
data <- data.frame(age = c(25, NA, 35), satisfaction = c("Satisfied", "Satisfied", NA))
          options <- list(missing = "fill", fill_value = "No Response", recode = TRUE,
                recode_rules = list(satisfaction = c("Not Satisfied" = "Dissatisfied")))
cleanSurveyData(data, options)
}
