% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurvey.R
\name{plotSurvey}
\alias{plotSurvey}
\title{Plot Survey}
\arguments{
\item{data}{A data frame containing survey data.}

\item{question}{A character string specifying the column name of the data to visualize.}

\item{type}{A character string indicating the type of visualization: "bar", "pie", or "histogram".}
}
\value{
A plot object displaying the chosen visualization.
}
\description{
Plot Survey
}
\examples{
data <- data.frame(satisfaction = c("Satisfied", "Dissatisfied", "Satisfied", "Neutral"))
          plotSurvey(data, question = "satisfaction", type = "bar")
}
